Implement a command runner module.
I seemed to have been in an OOP mood when I wrote the "plugin"
section of the README, and I realised that a much more functional
(and thus, in Haskell, easier) way to do it is to have partially-
applied state-consuming functions. This is how the command runner
is now implemented.

Sample commands have been added to Yukibot